fix(scroll): fallback to window.scrollTo
For some reason, `sendWheelEvent()` doesn't always actually trigger a scroll until
the user *actually* scrolls with a mouse/trackpad.

I've been able to ~somewhat reliably reproduce this by opening two tabs, rapidly
switching between them and then trying to scroll.

I couldn't figure out a *real* fix for this, so this is an attempt to workaround this bug
by checking if the window was scrolled within 5 frames, and scrolling using a separate method
that does seem to work when `sendWheelEvent()` does not.

Note: this entirely breaks nested frame scrolling, but as this bug is *very* annoying, that
       is a tradeoff we'll have to make for now.

Note: we can't just always use `.scrollTo()` because that *also* doesn't work in certain *other* cases.