fix(hints): revert svg change
there were too many false positives, I need to come up with a better
solution.

sometimes it would result in some elements not being clickable at all
because the svg label was in front of the actual label which is just
annoying. i'm not sure what the better heuristic would be though....