refactor(internal/hints): move browser viewport filtering to the content process
this is definitely less clean but I want to add a TS API for filtering /
mapping the returned hints, but then you can't filter accurately in the
content process because you'll be given hints that would never actually
be rendered.

then, I don't think it should be on the user to think about whether they
want to filter in the content process or not, i.e. we shouldn't have two
APIs, one for filtering in the content and one in the main thread. so
the only way we can fulfill that is by moving more logic to the content.