feat(commandline): add arrow keys as alternatives to tab
people be complaining that Tab and S-Tab is not intuitive which I guess
is kinda fair.

I originally didn't map it to arrow keys so that I could keep arrow keys
for navigating history. I think history will not be that important that
it would be worth the unintuitiveness of Tab