ci: only deploy docs from the docs branch
auto deploying from main is nice while testing but it means that the
docs can get out of sync with the actual released version.

now we only deploy when commits are pushed to a `docs` branch, so if a
docs change is release agnostic we can just cherry pick it to that
branch.

and when a release happens, we reset the docs branch to match the
release state