refactor(internal): move license check to internal script
I eventually want to completely move away from glider and just rely on
internal scripts to replace it, we only really use it for downloading
the firefox source, extracked packaged artifacts and  applying /
exporting patches and I'd much rather maintain that in-tree.

there's also a bunch of surprising behaviour in the current glider tree
like patching mozconfig files which makes grepping for the source of
said changes much more annoying.