fix(devtools): avoid crash under certain circumstances
sometimes when opening the browser toolbox or the repl, the dev tools
would crash, this happens because dev tools finds the fake <browser>
element that ends up in the document we pass to the config sandbox.

I think the real fix would be to never create this element in the first
place. Hoping to do that later.