fix(sandbox): avoid timer throttling
this binds timer functions to the chrome window instead of the sandbox
window to avoid cases where firefox throttles timers that are ran in a
"background context". our config sandbox window is treated as a
background context as its marked as hidden, if there are other issues
this causes then we may want to figure out how to mark it as not a
background context.

for more context see the big comment in
src/glide/browser/base/content/sandbox.mts