fix(sandbox/document): only mirror after the full browser document has loaded
previously, the mirror Document in the sandbox would be missing lots of
elements, this ensures every element in the browser UI is actually
copied over to the mirror.

this does now mean that any mutations directly in the config will not
work, so we report an error in that case.