feat(hint): add numeric label generator helper
The function is only a few lines, but we can help out users by providing
a stock implementation. Note that we ignore hint_chars here!

You might think that we could provide just a "sequential" generator that
would turn an alphabet like "abc" into:

  a, b, c, aa, ab, ac, ba, bb, bc, ca, cb, cc, aaa, ...

and then feed it the numeric digits via hint_chars. But that is not
quite how numbers work. If the alphabet is "1234567890", then we get "0"
instead of "10" as the tenth label, and "10" as the twentieth. If you
instead use "0123456789", then we start with "0" and get "01", "02",
etc, instead of the teens. You'd have to be smart about how zeroes are
handled, at which point it is easier to just have a custom generator
that simply counts.