feat(api/addons): only reinstall addons when necessary
we now cache the `glide.addons.install_from_url` calls so that if the
given XPI URL has not changed, then we'll skip reinstalling the addon.

this handles cases where the addon has been updated in Glide directly,
and the XPI URL no longer directly matches the version of the installed
addon. this is intentional, as I'm going to also rename this method to
make the functionality more clear that you're just "requesting" that the
addon corresponding to the given XPI URL has been installed - it's not a
lock file.

more reasoning here is that it should be very rare that you actually
want to use an older version of an extension.

we may introduce an actual addons lock file in the future, the glide API
call is not the right place to do so, as it is user-controlled.