feat(profile): store used versions
this will be very useful for changing default behaviour in the future,
as we can then display a notification and make it easy to revert back to
the previous behaviour if desired.

I spent a lot of time trying to figure out if Firefox already stored
this information somewhere, and I believe they do not. The closest I got
was an XML cache of the previously installed update files, but that
would only work for users going through the builtin auto-update flow.