feat(internal): store last used version in the profile directory
Firefox currently only stores the last used version in a
compatability.ini file, but that is written to very early in startup so
we can't refference that in our code as it'll just be the same as the
current version.

So this patch ensures we write a
`glide__compatibility_oldest_version.txt` file the first time Gide is
launched, so that we can actually reference the previously installed
version.

Note this is redundant with `.glide/versions.json` but we still need it
for any breaking changes made after 0.1.53a, as that file won't be
available.