feat(scrolling): make page scrolling work while an input element is focused
previously, if an <input> element was focused, these scrolls would do
almost nothing because they'd try to scroll the element instead of
the page.

the previous behaviour could make more sense for <textarea> because they
could actually have a scroll, but it does *not* make sense to support
that through the `scroll_page_up` etc excmds, as they explicitly
reference the page.

note that this probably applies to other kinds of elements too, not just
editable ones, but we can figure that out later.