refactor(internal): move glide API implementation to browser-api.mts
This is generally slightly cleaner, and also makes it easy to add lint
rules around the API implementation, for example there's currently a bug
where we'll always resolve relative paths relative to the config file,
even if we're evaluating a file from a completely different directory.

So the API being in a separate file makes it easier to lint what other
internal APIs it should access.