fix(extensions/api): properly serialise browser.contentScripts.register() (#186)
previously if you called `browser.contentScripts.register()` you'd get an error about a value that couldn't be cloned. this happened because it returned an object like `{ unregister() { ... } }` and functions cannot be cloned.

now we strip `unregister()` before cloning it, and add `$glide_content_functions` to the result object so that the main process knows that there was a function defined in the content process, which it will then add back in, and forward any calls back to the content process.

closes #174